consolidated classes and operations into a single python file
