+++ 02-dlt-meta-ingestion/classDefinitions.py
@@ -1,13 +1,59 @@
-# Databricks notebook source
-# MAGIC %run "./operations"
+import dlt
+from pyspark.sql.functions import *
+from pyspark.sql import DataFrame
+from pyspark.sql.session import SparkSession
+from pyspark.sql.streaming import DataStreamReader, DataStreamWriter
+from typing import Callable
 
-# COMMAND ----------
+##########################
+####### operations #######
+##########################
+
+# set the catalog and schema 
+def use_catalog_schema(catalog: str, schema: str, env_mode: str = "dev", verbose: bool = False):
+    if env_mode == "prd":
+        catalog_stmnt = f"""use catalog {catalog};"""
+    else:
+        catalog_stmnt = f"""use catalog {catalog}_{env_mode};"""
+    
+    spark.sql(catalog_stmnt)
+    spark.sql(f"""use schema {schema};""")
+    if verbose:
+        return spark.sql("""select current_catalog(), current_schema();""")
+
+# read streaming data as whole text using autoloader    
+def read_stream_raw(spark: SparkSession, path: str, maxFiles: int, maxBytes: str, wholeText: bool = True, options: dict = None) -> DataFrame:
+    stream_schema = "value STRING"
+    read_stream = (
+        spark
+        .readStream
+        .format("cloudFiles")
+        .option("cloudFiles.format", "text")
+        .option("wholetext", wholeText)
+        .option("cloudFiles.maxBytesPerTrigger", maxBytes)
+        .option("cloudFiles.maxFilesPerTrigger", maxFiles)
+    )
+
+    if options is not None:
+        read_stream = read_stream.options(**options)
+
+    read_stream = (
+        read_stream
+        .schema(stream_schema)
+        .load(path)
+    )
+
+    return read_stream
+
+###########################
+### classes and methods ###
+###########################
 
 class IngestionDLT:
 
     def __init__(
         self
-        ,spark: SparkSession = spark
+        ,spark: SparkSession # = spark
         ,env_mode: str = "dev"
         ,catalog: str = "lakehouse"
         ,schema: str = "landing"
@@ -71,4 +117,3 @@ class IngestionDLT:
         """
         for i in range(0,len(table_names)):
             ingest_raw_to_bronze(self = self, table_name = table_names[i], table_comment = table_comments[i], source_folder_path_from_volume = source_folder_path_from_volumes[i], table_properties = table_properties, maxFiles = maxFiles, maxBytes = maxBytes, wholeText = wholeText, options = options)
-
